+++ pkg/fs/filemgr.go
@@ -10,7 +10,7 @@ import (
 
 type FileMgr struct {
 	dbDirectory *os.File
-	blockSize   uint32
+	blockSize   int
 	files       map[string]*os.File
 	mu          *sync.Mutex
 }
@@ -29,7 +29,7 @@ func NewFileManager(directory string, blkSize int) (*FileMgr, error) {
 		return nil, err
 	}
 
-	return &FileMgr{dbDirectory: dir, blockSize: uint32(blkSize), files: make(map[string]*os.File), mu: &sync.Mutex{}}, nil
+	return &FileMgr{dbDirectory: dir, blockSize: blkSize, files: make(map[string]*os.File), mu: &sync.Mutex{}}, nil
 }
 
 func (f *FileMgr) Read(blk *BlockId, p *Page) {
@@ -86,16 +86,16 @@ func (f *FileMgr) Append(filename string) *BlockId {
 	return blk
 }
 
-func (f *FileMgr) Length(filename string) (int16, error) {
+func (f *FileMgr) Length(filename string) (int, error) {
 	file := f.getFile(filename)
 	stat, err := file.Stat()
 	if err != nil {
 		return 0, err
 	}
-	return int16(stat.Size()) / int16(f.blockSize), nil
+	return int(int32(stat.Size()) / int32(f.blockSize)), nil
 }
 
-func (f *FileMgr) BlockSize() uint32 {
+func (f *FileMgr) BlockSize() int {
 	return f.blockSize
 }
